Move packages/devhints-icons to src/web-icons/
