+++ src/web/components/HomeButton.tsx
@@ -1,6 +1,6 @@
-import { homeLine } from 'devhints-icons'
 import Link from 'gatsby-link'
 import React from 'react'
+import { homeLine } from '../../web-icons'
 import CSS from './HomeButton.module.scss'
 
 /**
